Cleanup test structure
* `Test.Common.*` in internal library `test-common`

  Two subhierarchies:
  - `Test.Common.HsBindgen.*` for `hs-bindgen`-specific functionality
  - `Test.Common.Util.*` for general purpose utilities

  Modules:
  - `Test.Common.HsBindgen.TracePredicate` was `Test.Internal.Tracer`
    (this does not provide tracer utils, but rather a _specific_ tracer).
  - `Test.Common.Util.Cabal` contains `findPackageDirectory`
  - `Test.Common.Util.Tasty.Golden` was `Test.Internal.TastyGolden`
  - `Test.Common.Util.Tasty` contains `assertException` and `goldenVsStringDiff_`

* Renamed the main test suite `test-internal` to `test-hs-bindgen`
  with corresponding module hierarchy `Test.HsBindgen.*`

  Unit tests in `Test.HsBindgen.Unit.*`:

  - `Test.HsBindgen.Unit.Tracer`
    (was `Test.HsBindgen.Util.Tracer`, we are now reserving `Util` for utilities
    for writing tests, rather than tests _of_ utils; see below).
  - `Test.HsBindgen.Unit.ClangArgs`
    Moved `testGetTargetTriple` here also (was `Test.HsBindgen.C.Parser`).

  Golden tests in `Test.HsBindgen.Golden.*`:

  - `Test.HsBindgen.Golden` currently contains all the golden tests, except for
    TH. This badly needs cleaning up, but will do that in the next PR.
  - `Test.HsBindgen.Golden.TH` for the golden TH tests
    (was `Test.Internal.TH`)

  Then some utilities that are specific to this test-suite (perhaps some of
  these could move to `test-common`):

  - `Test.HsBindgen.Util.Clang` for setting up the clang environment
  - `Test.HsBindgen.Util.Rust` for calling `rust-bindgen`

  Finally we have

  - `Test.HsBindgen.Orphans.TreeDiff` for the `ToExpr` orphans

* `Test.TH.*` for `test-th`, and similarly `Test.PP.*` for `test-pp`
